Fixed module import to load plugins remote to pkg.
Fixed the _loadModule function in the custom analyser to correctly load a
module using importlib.

Added function override to Locator.locatePlugins so the candidate full
path is given to the analyser, not just the name.  This is required for
the analyser to correctly load a module.  Mostly the function is copied
from the equivalent in Yapsy PluginFileLocator.

Minor fixes to the test data to make the tests pass (they were checking
the wrong things before).